Helper for get file type implemented
